+++ zcash_client_backend/CHANGELOG.md
@@ -18,6 +18,7 @@ and this library adheres to Rust's notion of
   - `NullifierQuery` for use with `WalletRead::get_sapling_nullifiers`
   - `BlockMetadata`
   - `ScannedBlock`
+  - `chain::CommitmentTreeRoot`
   - `wallet::input_sellection::Proposal::{min_target_height, min_anchor_height}`:
 - `zcash_client_backend::wallet::WalletSaplingOutput::note_commitment_tree_position`
 - `zcash_client_backend::scanning::ScanError`